integrate sorted set without view - rename collections folder to not conflict with built in collections
