+++ edgegraph/structure/collections_/__init__.py
@@ -2,6 +2,6 @@
 Holds the custom collections unique to edgegraph.
 """
 
-from edgegraph.structure.collections.sortedset import SortedSet
+from edgegraph.structure.collections_.sortedset import SortedSet
 
 __all__ = ["SortedSet"]
