+++ edgegraph/structure/universe.py
@@ -10,6 +10,7 @@ import threading
 from typing import TYPE_CHECKING
 
 from edgegraph.structure import vertex
+from edgegraph.structure.collections_.sortedset import SortedSet
 
 if TYPE_CHECKING:
     Vertex = vertex.Vertex
@@ -59,7 +60,7 @@ class Universe(vertex.Vertex):
         self._verts_lock = threading._PyRLock()
 
         #: Internal set of vertices
-        self._vertices: list[Vertex] = []
+        self._vertices: SortedSet[Vertex] = SortedSet()
         if vertices is not None:
             for v in vertices:
                 self.add_vertex(v)
@@ -102,7 +103,7 @@ class Universe(vertex.Vertex):
             if vert in self._vertices:
                 return
 
-            self._vertices.append(vert)
+            self._vertices.add(vert)
             if self not in vert.universes:
                 vert.add_to_universe(self)
